extract server health contract types into core microcrate
