+++ Cargo.toml
@@ -41,6 +41,7 @@ members = [
   "crates/bitnet-gguf",          # SRP: lightweight GGUF file format types & header parser
   "crates/bitnet-eval-core",     # SRP: deterministic eval/scoring math helpers
   "crates/bitnet-feature-matrix",
+  "crates/bitnet-server-health-types-core", # SRP: reusable server health endpoint contract types
   "crates/bitnet-server",
   "crates/bitnet-cli-sampling-core", # SRP: reusable CLI sampling primitives
   "crates/bitnet-cli",
@@ -136,6 +137,7 @@ default-members = [
   "crates/bitnet-bdd-grid-core",
   "crates/bitnet-feature-contract",
   "crates/bitnet-feature-matrix",
+  "crates/bitnet-server-health-types-core",
   "crates/bitnet-testing-policy-core",
   "crates/bitnet-testing-scenarios-core",
   "crates/bitnet-runtime-profile-core",
