+++ crates/bitnet-server-health-types-core/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "bitnet-server-health-types-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "Reusable health endpoint response contracts for BitNet server"
+rust-version.workspace = true
+
+[dependencies]
+serde.workspace = true
+
+[dev-dependencies]
+serde_json.workspace = true
+
+[lints]
+workspace = true
